pyproject.toml
[tool.poetry]
name = "language_detection"
version = "0.1.0"
description = ""
authors = ["ajms <albert.schulz@posteo.de>"]
license = "GPL3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
bentoml = "^1.1.10"
transformers = "^4.35.2"
torch = "^2.1.1"
wheel = "^0.42.0"
setuptools = "^69.0.2"
pip = "^23.3.1"
pybind11 = "^2.11.1"
fasttext-wheel = "^0.9.2"


[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
flake8 = "^6.1.0"
isort = "^5.12.0"
pre-commit-hooks = "^4.5.0"
pre-commit = "^3.5.0"
ipykernel = "^6.27.1"
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
exclude = '''
/(
    \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | _build
    | buck-out
    | build
    | dist
)/
'''
include = '\.pyi?$'
line-length = 88

[tool.isort]
include_trailing_comma = true
line_length = 88
multi_line_output = 3
profile = "black"
skip_gitignore = true
